% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme.R
\name{meso_scheme_rules}
\alias{meso_scheme_rules}
\title{Extract the rules from a mesohabitat scheme}
\usage{
meso_scheme_rules(x)
}
\arguments{
\item{x}{A \code{meso_scheme}.}
}
\value{
A plain data frame containing the rule table.
}
\description{
Extract the rules from a mesohabitat scheme
}
\examples{
meso_scheme_rules(meso_scheme_default())
}
