% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{mesohabitat_example_vector}
\alias{mesohabitat_example_vector}
\title{Create a synthetic example spatial vector}
\usage{
mesohabitat_example_vector()
}
\value{
A point \code{terra::SpatVector} in EPSG:32615 made from
\link{hydromeso_example}.
}
\description{
Create a synthetic example spatial vector
}
\examples{
mesohabitat_example_vector()
}
